requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".
fastapi
google-cloud-storage
redis
uvicorn[standard]
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-google-cloud

